Refactored fetching items to maintain DRY
